requirements.txt
asttokens==2.0.5
backcall==0.2.0
backports.functools-lru-cache==1.6.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cox==0.1.post3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
dill==0.3.5.1
docker-pycreds==0.4.0
einops==0.4.1
entrypoints==0.4
executing==0.8.3
fonttools==4.30.0
gitdb==4.0.9
GitPython==3.1.27
GPUtil==1.4.0
grpcio==1.48.0
idna==3.3
ipykernel==6.9.1
ipython==8.1.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.2
lmdb==1.3.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numexpr==2.8.3
numpy==1.22.3
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
promise==2.3
prompt-toolkit==3.0.27
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py3nvml==0.2.7
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==19.0.2
requests==2.27.1
robustness==1.2.1.post2
scikit-learn==1.1.1
scipy==1.8.0
seaborn==0.11.2
sentry-sdk==1.5.9
setproctitle==1.2.3
setuptools==58.0.4
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
tables==3.7.0
tensorboardX==2.5.1
threadpoolctl==3.1.0
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
wandb==0.12.14
wcwidth==0.2.5
wheel==0.37.1
xmltodict==0.13.0
